Abstract Camera duplication + rendering into CameraCapture
Moving the logic for duplicating cameras and rendering a screenshot into
a base class gives a common ground for the different camera types and
makes it easier to write a single class to manage them.
