+++ Telemachus/Telemachus.csproj
@@ -65,6 +65,7 @@
       <DesignTime>True</DesignTime>
       <DependentUpon>Resources.resx</DependentUpon>
     </Compile>
+    <Compile Include="src\CameraSnapshots\CameraCapture.cs" />
     <Compile Include="src\CameraSnapshots\IVACameraCapture.cs" />
     <Compile Include="src\CameraSnapshots\CurrentFlightCameraCapture.cs" />
     <Compile Include="src\CameraSnapshots\RasterPropMonitorCamera.cs" />
