+++ Telemachus/src/CameraSnapshots/CurrentFlightCameraCapture.cs
@@ -5,102 +5,15 @@ using System.Collections.Generic;
 
 namespace Telemachus.CameraSnapshots
 {
-    public class CurrentFlightCameraCapture : MonoBehaviour
+    public class CurrentFlightCameraCapture : CameraCapture
     {
-        public RenderTexture overviewTexture;
-        public bool didRender;
-        public byte[] imageBytes = null;
-        public bool mutex = false;
+        protected static new string cameraContainerNamePrefix = "Telemachus Camera Container - ";
 
-        protected Dictionary<string, Camera> cameraDuplicates = new Dictionary<string, Camera>();
-        protected List<string> activeCameras;
-        protected static readonly string[] skippedCameras = { "UIMainCamera", "UIVectorCamera" };
-        protected static string cameraContainerNamePrefix = "Telemachus Camera Container - ";
-        
-        private const float fovAngle = 60f;
-        private const float aspect = 1.0f;
-        private  int camerares = 300;
-
-        void LateUpdate()
+        public override void BeforeRenderNewScreenshot()
         {
-            PluginLogger.debug("LATEUPDATE FOR CAMERA");
-            if(CameraManager.Instance != null)
-            {
-                PluginLogger.debug("CAMERA FOUND, TAKING SCREENSHOT");
-                StartCoroutine(NewScreenshot());
-            }
-        }
-
-        public void UpdateCameras()
-        {
-            if (CameraManager.Instance != null)
-            {
-                PluginLogger.debug("CURRENT CAMERA MODE: " + CameraManager.Instance.currentCameraMode);
-            }
-
-            activeCameras = new List<string>();
-
-            foreach (Camera camera in Camera.allCameras)
-            {
-                debugCameraDetails(camera);
-                // Don't duplicate any cameras we're going to skip
-                if (skippedCameras.IndexOf(camera.name) != -1)
-                {
-                    continue;
-                }
-
-                Camera cameraDuplicate;
-
-                if (!cameraDuplicates.ContainsKey(camera.name))
-                {
-                    var cameraDuplicateGameObject = new GameObject(cameraContainerNamePrefix + camera.name);
-                    cameraDuplicate = cameraDuplicateGameObject.AddComponent<Camera>();
-                    cameraDuplicates[camera.name] = cameraDuplicate;
-                }
-                else
-                {
-                    cameraDuplicate = cameraDuplicates[camera.name];
-                }
-
-                cameraDuplicate.CopyFrom(camera);
-                cameraDuplicate.enabled = false;
-                cameraDuplicate.fieldOfView = fovAngle;
-                cameraDuplicate.aspect = aspect;
-
-                //Now that the camera has been duplicated, add it to the list of active cameras
-                activeCameras.Add(camera.name);
-            }
-        }
-
-        public void debugCameraDetails(Camera cam)
-        {
-            PluginLogger.debug("CAMERA: " + cam.name + " ; FAR: " + cam.far + "; FAR CLIP PLANE: " + cam.farClipPlane);
-        }
-        
-        public IEnumerator NewScreenshot()
-        {
-            if (mutex)
-            {
-                yield return true;
-            }
-            PluginLogger.debug("BYPASSED MUTEX");
-            mutex = true;
-            PluginLogger.debug("WAITING FOR END OF FRAME");
-            yield return new WaitForEndOfFrame();
-
             UpdateCameras();
-
-            List<Camera> renderingCameras = new List<Camera>();
-            foreach (string cameraName in activeCameras)
-            {
-                PluginLogger.debug("GETTING CAMERA" + cameraName);
-                renderingCameras.Add(cameraDuplicates[cameraName]);
-            }
-
-            this.imageBytes = SnapshotRenderer.renderSnaphot(renderingCameras, camerares, camerares);
-            this.didRender = true;
-            mutex = false;
+            base.BeforeRenderNewScreenshot();
         }
-
+        
     }
 }
